updates teamRoutes, homeRoutes, teampage.js and joinTeam handlebars to allow users to create a team, join a created team and still trying to fix rendering all users teams on joinTeam page
