+++ controllers/homeRoutes.js
@@ -19,8 +19,23 @@ router.get('/signup', async (req, res) => {
 });
 
 router.get('/joinTeam', Authenticated, async (req, res) => {
-
-    res.render('joinTeam');
+    try {
+        const userTeamData = await User.findByPk(req.session.user_id, {
+            attributes: {exclude: ['password', 'first_name', 'last_name', 'email']},
+            include:[{model:Team, attributes: {exclude: ['password']}}],
+        });
+        if(!userTeamData) {
+            res.render('joinTeam');
+        }
+        const teams = userTeamData.teams
+        
+        const plainTeams = teams.map((teams) => teams.get({plain:true}));
+        res.render('joinTeam', {
+            plainTeams
+        });
+    } catch {
+        res.render('joinTeam');
+    }
 });
 
 router.get('/teamDash', Authenticated, async (req, res) => {
