+++ views/joinTeam.handlebars
@@ -53,6 +53,7 @@
                         <button class="btn btn-outline-light btn-lg px-5" type="submit">Team 3</button> --}}
                         {{#each plainTeam as |team| }}
                         <button id="{{team.id}}" class="btn btn-outline-light btn-lg px-5" type="submit">{{team.name}}</button> 
+                        {{/each}}
                     </div>
                     <div class="d-flex justify-content-center text-center mt-4 pt-1">
                       <a href="#!" class="text-white"><i class="fab fa-facebook-f fa-lg"></i></a>
